Обновлён ICR

+++ UsefulMFC/ClrButton.cpp
@@ -235,6 +235,6 @@ void CClrButton::DrawButtonText(CDC *dc, CRect r, const char *Buf, COLORREF text
 
 	previous_colour = dc->SetTextColor(text_colour);
     dc->SetBkMode(TRANSPARENT);
-	dc->DrawText(Buf, strlen(Buf), r, DT_CENTER | DT_VCENTER | DT_SINGLELINE);
+	dc->DrawText(Buf, (int)strlen(Buf), r, DT_CENTER | DT_VCENTER | DT_SINGLELINE);
 	dc->SetTextColor(previous_colour);
 }
